MDEV-23082: ER_TABLEACCESS_DENIED_ERROR error message is truncated, and inaccurately
The list of privilege is 110 characters wide. However, the format of error
message was: "%-.32s command denied to user '%s'@'%s' for table '%-.192s'".
get_length() sets the maximum width to 32 characters. Hence, eventually,
only first 32 characters of list of privilege are stored.
Fix: Changing the format to "%-.110..." so that get_length() sets width to 110.
Hence, all 110 characters of the list of privilege are stored.